requirements.txt
flask==2.2.2
flasgger==0.9.5
psycopg2-binary==2.9.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
pytest==6.2.5
python-dotenv==0.21.0
Flask-Cors==3.0.10
requests==2.28.1
flask-caching==2.0.1
PyJWT==2.3.0
marshmallow==3.16.0
marshmallow-enum==1.5.1
flask-bcrypt==1.0.1
dependency_injector==4.38.0
